Add tool-versions to force Erlang/Elixir versions
The goal is to make it less easy to build the quickstart with old
or incompatible versions of Erlang and Elixir. Many other versions
should work, but since this is a starter project for many new Nerves
users, it just seems easier to help them debug if we're using the same
tools.