Use board-specific network configurations
Previously a generic setup was used that worked across all boards
with the caveat that it would configure some impossible-to-use
interfaces on some boards.

The NPI imx6ull board has an additional Ethernet interface. The two
options for supporting it are to add it to the generic config or to
split out the config per board. The latter option seems like it will
lead to the least confusion so go with it.