Don't set the hostname pattern in the config
This repeats what's already being set in the official Nerves systems.
When using this project on unofficial systems, it is a little weird to
override the hostnames.

Since examples are nice, this changes it to set the time on boot. This
gets the date a little closer to the real date and avoids all of the
1970 time logs on init.